Shifting click_options to cli dir and init_config to config handler
