+++ calm/dsl/cli/main.py
@@ -11,12 +11,12 @@ from prettytable import PrettyTable
 # TODO - move providers to separate file
 from calm.dsl.providers import get_provider, get_provider_types
 from calm.dsl.api import get_api_client, get_resource_api
-from calm.dsl.tools import simple_verbosity_option, show_trace_option
 from calm.dsl.log import get_logging_handle
 from calm.dsl.config import get_context
 from calm.dsl.store import Cache
 
 from .version_validator import validate_version
+from .click_options import simple_verbosity_option, show_trace_option
 from .utils import FeatureFlagGroup, highlight_text
 
 CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
