+++ calm/dsl/tools/__init__.py
@@ -1,6 +1,5 @@
 from .ping import ping
 from .validator import StrictDraft7Validator
-from .click_options import simple_verbosity_option, show_trace_option
 from .utils import get_module_from_file, make_file_dir
 
 
@@ -8,8 +7,6 @@ __all__ = [
     "RenderJSON",
     "ping",
     "StrictDraft7Validator",
-    "simple_verbosity_option",
-    "show_trace_option",
     "get_module_from_file",
     "make_file_dir",
 ]
